[Install] Fixed column so it stores double for time

+++ install.php
@@ -70,7 +70,7 @@ if(isset($_POST))
 		                `username` varchar(255) NOT NULL,
 		                `serverId` varchar(255) NOT NULL,
 		                `ip` varchar(50) NOT NULL,
-		                `time` bigint(11) UNSIGNED NOT NULL,
+		                `time` DOUBLE(40,2) UNSIGNED NOT NULL,
 		                PRIMARY KEY (`id`)
 		            ) ENGINE=InnoDB DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;
